#86 chore : 패캐지 네이밍 snake_case -> nocase 변경
