+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/edit/EditProfileBindingAdapter.kt
@@ -1,4 +1,4 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage.edit_profile
+package com.avengers.nibobnebob.presentation.ui.main.mypage.edit
 
 import android.util.Log
 import android.widget.TextView
