+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/wishlist/WishRestaurantListFragment.kt
@@ -1,4 +1,4 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage.my_list
+package com.avengers.nibobnebob.presentation.ui.main.mypage.wishlist
 
 import android.os.Bundle
 import android.view.View
@@ -7,18 +7,18 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
 import com.avengers.nibobnebob.R
-import com.avengers.nibobnebob.databinding.FragmentMyRestaurantListBinding
+import com.avengers.nibobnebob.databinding.FragmentWishRestaurantListBinding
 import com.avengers.nibobnebob.presentation.base.BaseFragment
 import com.avengers.nibobnebob.presentation.ui.main.MainViewModel
 import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedUiEvent
 import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedViewModel
 
-class MyRestaurantListFragment :
-    BaseFragment<FragmentMyRestaurantListBinding>(R.layout.fragment_my_restaurant_list) {
+class WishRestaurantListFragment :
+    BaseFragment<FragmentWishRestaurantListBinding>(R.layout.fragment_wish_restaurant_list) {
 
+    override val parentViewModel: MainViewModel by activityViewModels()
     private val sharedViewModel: MyPageSharedViewModel by viewModels()
     private lateinit var navController: NavController
-    override val parentViewModel: MainViewModel by activityViewModels()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -36,7 +36,7 @@ class MyRestaurantListFragment :
             sharedViewModel.uiEvent.collect { event ->
                 when (event) {
                     is MyPageSharedUiEvent.NavigateToBack ->
-                        navController.navigate(MyRestaurantListFragmentDirections.globalToMyPageFragment())
+                        navController.navigate(WishRestaurantListFragmentDirections.globalToMyPageFragment())
 
                     else -> Unit
                 }
