+++ Aos/app/src/main/res/layout/fragment_edit_profile.xml
@@ -9,11 +9,11 @@
 
         <variable
             name="ui_state"
-            type="com.avengers.nibobnebob.presentation.ui.main.mypage.edit_profile.EditProfileUiState" />
+            type="com.avengers.nibobnebob.presentation.ui.main.mypage.edit.EditProfileUiState" />
 
         <variable
             name="vm"
-            type="com.avengers.nibobnebob.presentation.ui.main.mypage.edit_profile.EditProfileViewModel" />
+            type="com.avengers.nibobnebob.presentation.ui.main.mypage.edit.EditProfileViewModel" />
 
         <variable
             name="svm"
@@ -23,7 +23,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.main.mypage.edit_profile.EditProfileFragment">
+        tools:context=".presentation.ui.main.mypage.edit.EditProfileFragment">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/gl_start"
